refactor(database): Convert NUMERIC data types to REAL
This is so that I don't have to manually convert Doubles, which get
returned as strings, to javascript Numbers from
the database. I think Strava only gives floats so this should be fine
for now

+++ packages/database/sql/003-athlete.sql
@@ -20,7 +20,7 @@ CREATE TABLE athletes (
     friend_count INT,
     measurement_preference TEXT,
     ftp INT,
-    weight NUMERIC
+    weight REAL
 );
 
 COMMIT;
